Move versionGreaterThan to standalone file.
