+++ internal/conf/reposource/jvm_packages.go
@@ -4,9 +4,7 @@ import (
 	"fmt"
 	"net/url"
 	"sort"
-	"strconv"
 	"strings"
-	"unicode"
 
 	"github.com/sourcegraph/sourcegraph/internal/api"
 )
@@ -135,58 +133,4 @@ func jdkModule() MavenModule {
 		GroupID:    "jdk",
 		ArtifactID: "jdk",
 	}
-}
-
-// versionGreaterThan return true if the version string a is "greater" than the
-// version string b using psuedo semantic versioning. Java package versions can
-// be arbitrary strings so this method must always succeed even if a version is
-// not using the semantic version format. The comparison is lexicographical by
-// default except for the digit parts which are compared numerically. For
-// example, versionGreaterThan return true when comparing 11.2.0 and 2.2.0
-// because the number 11 is larger than 2 (even if "2" is lexicographically
-// larger than "11").
-func versionGreaterThan(version1, version2 string) bool {
-	index := 0
-	end := len(version1)
-	if len(version2) < end {
-		end = len(version2)
-	}
-	for index < end {
-		rune1 := rune(version1[index])
-		rune2 := rune(version2[index])
-		if unicode.IsDigit(rune1) && unicode.IsDigit(rune2) {
-			int1 := versionParseInt(index, version1)
-			int2 := versionParseInt(index, version2)
-			if int1 == int2 {
-				index = versionNextNonDigitOffset(index, version1)
-			} else {
-				return int1 > int2
-			}
-		} else {
-			if rune1 == rune2 {
-				index += 1
-			} else {
-				return rune1 > rune2
-			}
-		}
-	}
-	return len(version1) < len(version2)
-}
-
-// versionParseInt returns the integer value of the number that appears at given
-// index of the given string.
-func versionParseInt(index int, a string) int {
-	end := versionNextNonDigitOffset(index, a)
-	value, _ := strconv.Atoi(a[index:end])
-	return value
-}
-
-// versionNextNonDigitOffset returns the offset of the next non-digit character
-// of the given string starting at the given index.
-func versionNextNonDigitOffset(index int, b string) int {
-	offset := index
-	for offset < len(b) && unicode.IsDigit(rune(b[offset])) {
-		offset += 1
-	}
-	return offset
 }
